Fix more PR comments
