+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:networkSecurityConfig="@xml/network_security_config"
         android:theme="@style/AppTheme">
 
-        <activity android:name="co.nimblehq.ui.screens.onboarding.OnboardingActivity"
+        <activity android:name="co.nimblehq.ui.screen.onboarding.OnboardingActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustNothing">
             <intent-filter>
@@ -25,7 +25,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="co.nimblehq.ui.screens.main.MainActivity"
+        <activity android:name="co.nimblehq.ui.screen.main.MainActivity"
             android:launchMode="singleTask"
             android:windowSoftInputMode="adjustNothing">
         </activity>
