+++ app/src/main/java/co/nimblehq/di/modules/ui/main/MainActivityModule.kt
@@ -1,8 +1,8 @@
 package co.nimblehq.di.modules.ui.main
 
 import android.app.Activity
-import co.nimblehq.ui.screens.main.MainNavigator
-import co.nimblehq.ui.screens.main.MainNavigatorImpl
+import co.nimblehq.ui.screen.main.MainNavigator
+import co.nimblehq.ui.screen.main.MainNavigatorImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
