+++ app/src/main/java/co/nimblehq/extension/ThrowableExtension.kt
@@ -4,5 +4,5 @@ import android.content.Context
 import co.nimblehq.R
 
 fun Throwable.userReadableMessage(context: Context): String {
-    return context.getString(R.string.error_generic) // TODO: Add more error messages handler here
+    return context.getString(R.string.generic_error) // TODO: Add more error messages handler here
 }
