+++ app/src/main/java/co/nimblehq/ui/screen/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.ui.screens.main
+package co.nimblehq.ui.screen.main
 
 import androidx.hilt.lifecycle.ViewModelInject
 import co.nimblehq.ui.base.BaseViewModel
