+++ app/src/main/java/co/nimblehq/ui/screen/onboarding/OnboardingActivity.kt
@@ -1,11 +1,11 @@
-package co.nimblehq.ui.screens.onboarding
+package co.nimblehq.ui.screen.onboarding
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import co.nimblehq.R
 import co.nimblehq.extension.blurView
 import co.nimblehq.ui.base.BaseActivity
-import co.nimblehq.ui.screens.onboarding.signin.BlurAnimatable
+import co.nimblehq.ui.screen.onboarding.signin.BlurAnimatable
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.android.synthetic.main.activity_onboarding.*
