+++ app/src/main/res/drawable/bg_user_login_button.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <corners android:radius="@dimen/_12sdp"/>
-            <solid android:color="@color/colorWhite"/>
-        </shape>
-    </item>
-</layer-list>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="@dimen/_12sdp"/>
+    <solid android:color="@color/colorWhite"/>
+</shape>
