+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/clMain"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="co.nimblehq.ui.screens.main.MainActivity">
+    tools:context="co.nimblehq.ui.screen.main.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fcvMainNavHostContainer"
