+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -6,6 +6,6 @@
     app:startDestination="@id/surveysFragment">
     <fragment
         android:id="@+id/surveysFragment"
-        android:name="co.nimblehq.ui.screens.main.surveys.SurveysFragment"
+        android:name="co.nimblehq.ui.screen.main.surveys.SurveysFragment"
         android:label="SurveysFragment" />
 </navigation>
