+++ app/src/main/res/values/strings_no_translation.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="sign_in_nimble_logo_desc">Nimble Logo</string>
+    <string name="onboarding_splash_background_desc">Splash Screen Background</string>
+</resources>
