+++ app/src/main/res/values/styles.xml
@@ -12,10 +12,10 @@
 
     <style name="LoginEditTextStyle">
         <item name="android:background">@drawable/bg_user_login_input</item>
-        <item name="android:singleLine">true</item>
+        <item name="android:maxLines">1</item>
         <item name="android:textColor">@color/colorWhite</item>
         <item name="android:textColorHighlight">@color/colorWhite</item>
-        <item name="android:textColorHint">@color/colorWhite</item>
+        <item name="android:textColorHint">@color/white_30a</item>
         <item name="android:textColorLink">@color/colorPrimaryDark</item>
         <item name="android:textCursorDrawable">@color/colorWhite</item>
         <item name="android:textSize">@dimen/_13ssp</item>
